moving mvn plugins to parent pom
